# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  app.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  app.CreditRequestInput:
    properties:
      amount:
        type: number
      memo:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - amount
    - memo
    type: object
  app.CreditUserResponse:
    properties:
      amount:
        type: number
      memo:
        type: string
      user:
        type: string
    type: object
...
  app.GenerateVouchersInput:
    properties:
      count:
        type: integer
      expire_after_days:
        type: integer
      value:
        type: number
    required:
    - count
    - expire_after_days
    - value
    type: object
  app.GetUserResponse:
    properties:
      account_address:
        type: string
      admin:
        type: boolean
      balance_in_tft:
        type: number
      code:
        type: integer
      credit_card_balance:
        description: millicent, money from credit card
        type: integer
      credit_card_balance_in_usd:
        type: number
      credited_balance:
        description: millicent, manually added by admin or from vouchers
        type: integer
      credited_balance_in_usd:
        type: number
      debt:
        description: millicent
        type: integer
      debt_in_usd:
        type: number
      email:
        type: string
      id:
        type: integer
      password:
        items:
          type: integer
        type: array
      sponsored:
        type: boolean
      ssh_key:
        type: string
      stripe_customer_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
      verified:
        type: boolean
    required:
    - email
    - password
    - username
    type: object
...
  app.SendMailResponse:
    properties:
      failed_emails:
        items:
          type: string
        type: array
      failed_emails_count:
        type: integer
      successful_emails:
        type: integer
      total_users:
        type: integer
    type: object
...
  app.TransferRecordsResponse:
    properties:
      created_at:
        type: string
      failure:
        type: string
      id:
        type: integer
      operation:
        $ref: '#/definitions/models.operation'
      state:
        $ref: '#/definitions/models.state'
      tft_amount:
        description: TFTs are multiplied by 1e7
        type: integer
      tft_amount_in_whole_unit:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  models.Voucher:
    properties:
      code:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      redeemed:
        type: boolean
      value:
        type: number
    required:
    - code
    - created_at
    - expires_at
    - value
    type: object
  models.operation:
    enum:
    - withdraw
    - deposit
    type: string
    x-enum-varnames:
    - WithdrawOperation
    - DepositOperation
  models.state:
    enum:
    - failed
    - success
    - pending
    type: string
    x-enum-varnames:
    - FailedState
    - SuccessState
    - PendingState
...
info:
  contact: {}
  description: API documentation for Mycelium Cloud.
  title: Mycelium Cloud API
  version: "1.0"
paths:
...
  /transfer-records:
    get:
      consumes:
      - application/json
      description: Returns all transfer records in the system
      operationId: list-transfer-records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/app.TransferRecordsResponse'
              type: array
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: List transfer records
      tags:
      - admin
...
  /users:
    get:
      consumes:
      - application/json
      description: Returns a list of all users
      operationId: get-all-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.GetUserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Get all users
      tags:
      - admin
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user from the system
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.APIResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/app.APIResponse'
        "403":
          description: Admins cannot delete their own account
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Delete a user
      tags:
      - admin
  /users/{user_id}/credit:
    post:
      consumes:
      - application/json
      description: Credits a specific user's balance
      operationId: credit-user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Credit Request Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app.CreditRequestInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.CreditUserResponse'
              type: object
        "400":
          description: Invalid request format or user ID
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Credit user balance
      tags:
      - admin
  /users/mail:
    post:
      consumes:
      - multipart/form-data
      description: Allows admin to send a custom email to all users with optional
        file attachments. Returns detailed statistics about successful and failed
        email deliveries.
      operationId: admin-mail-all-users
      parameters:
      - description: Email subject
        in: formData
        name: subject
        required: true
        type: string
      - description: Email body content
        in: formData
        name: body
        required: true
        type: string
      - description: Email attachments (multiple files allowed)
        in: formData
        name: attachments
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Email sending results with delivery statistics
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/app.SendMailResponse'
              type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Send mail to all users
      tags:
      - admin
  /vouchers:
    get:
      consumes:
      - application/json
      description: Returns all vouchers in the system
      operationId: list-vouchers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Voucher'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: List vouchers
      tags:
      - admin
  /vouchers/generate:
    post:
      consumes:
      - application/json
      description: Generates a bulk of vouchers
      operationId: generate-vouchers
      parameters:
      - description: Generate Vouchers Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app.GenerateVouchersInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/app.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Voucher'
                  type: array
              type: object
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/app.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.APIResponse'
      security:
      - AdminMiddleware: []
      summary: Generate vouchers
      tags:
      - admin
...
swagger: "2.0"